RGB Pixel Manipulation
This program separates pixel values of an RGB image into three bands,
R,G, and B, separately. It then displays those images individually with
a black overlay, only displaying a small portion of the screen at a
time. The point of this is to deconstruct an image, and reconstruct it
using a long exposure with a DSLR. This is a proof of concept for Bayer
filters, additive light, and long exposure photography.